\docType{package}


\name{lmSubsets-package}


\alias{lmSubsets-package}


\title{Package \code{lmSubsets}}


\description{

  Variable-subset selection in ordinary linear regression.

}


\author{

  \itemize{

    \item{Marc Hofmann (\email{marc.hofmann@gmail.com})}

    \item{Cristian Gatu (\email{cgatu@info.uaic.ro})}

    \item{Erricos J. Kontoghiorghes (\email{erricos@dcs.bbk.ac.uk})}

    \item{Ana Colubi (\email{ana.colubi@gmail.com})}

    \item{Achim Zeileis (\email{Achim.Zeileis@R-project.org})}

  }

}


\references{

  Hofmann M, Gatu C, Kontoghiorghes EJ, Colubi A, Zeileis A (2020).
  lmSubsets: Exact variable-subset selection in linear regression for R.
  \emph{Journal of Statistical Software}, \strong{93}, 1--21.
  \doi{10.18637/jss.v093.i03}.

  Hofmann M, Gatu C, Kontoghiorghes EJ (2007).  Efficient algorithms for
  computing the best subset regression models for large-scale problems.
  \emph{Computational Statistics & Data Analysis}, \strong{52}, 16--29.
  \doi{10.1016/j.csda.2007.03.017}.

  Gatu C, Kontoghiorghes EJ (2006).  Branch-and-bound algorithms for
  computing the best subset regression models.  \emph{Journal of
  Computational and Graphical Statistics}, \strong{15}, 139--156.
  \doi{10.1198/106186006x100290}.

}


\seealso{

  Home page:  \url{https://github.com/marc-hofmann/lmSubsets.R}

}
