\name{summary.lmSubsets}


\alias{summary.lmSubsets}
\alias{summary.lmSelect}
\alias{print.summary.lmSubsets}
\alias{print.summary.lmSelect}


\title{Summarize a subset regression}


\description{

  Evaluate summary statistics for the selected submodels.

}


\usage{
\method{summary}{lmSubsets}(object, ..., na.rm = TRUE)

\method{summary}{lmSelect}(object, ..., na.rm = TRUE)
}


\arguments{

  \item{object}{\code{"lmSubsets"}, \code{"lmSelect"}---a subset
    regression}

  \item{...}{ignored}

  \item{na.rm}{if \code{TRUE}, remove \code{NA} values}

}


\value{

  \code{"summary.lmSubsets"}, \code{"summary.lmSelect"}---a subset
  regression summary

}


\seealso{

  \itemize{

    \item{\code{\link[=lmSubsets]{lmSubsets()}} for all-subsets
      regression}

    \item{\code{\link[=lmSelect]{lmSelect()}} for best-subset
      regression}

  }

}
